Bug 10662: Build OAI-PMH Harvesting Client
This patch adds an OAI-PMH harvesting client to Koha.

The client runs as a daemon in the background. Users interact with the client
via the Koha web user interface, which communicates with the daemon via a unix socket
using a simple JSON-based protocol.

The harvester ingests MARCXML. You can harvest other metadata formats, but you
must use a XSLT to transform them into MARCXML, if you want them to be imported
into Koha.

You can supply your own download and import modules via the oai-pmh-harvester.yaml
configuration file, but the default modules supplied in this patch should
be good enough for your purposes. If they're not, raise a Bugzilla issue.

There is a cleanup_database.pl addition, because high volume harvesting
will cause the oai_harvester_import_queue table to fill quickly. This table
is not required for adding/updating records. It's mostly just for general
monitoring and audit purposes.

Signed-off-by: Andreas Hedström Mace <andreas.hedstrom.mace@sub.su.se>